feat(math): scaleToContain/scaleToCover added, scaleToFit deprecated
