+++ src/math/mod.ts
@@ -1,12 +1,14 @@
-export type * from './types.ts';
 export * from './clamp.ts';
 export * from './deg-to-rad.ts';
+export * from './get-numbers-around.ts';
+export * from './get-pseudo-random.ts';
 export * from './lerp.ts';
 export * from './rad-to-deg.ts';
 export * from './random-float.ts';
 export * from './random-integer.ts';
+export * from './scale-rect-to-contain.ts';
+export * from './scale-rect-to-cover.ts';
 export * from './scale-to-fit.ts';
 export * from './shuffle.ts';
 export * from './vector2.ts';
-export * from './get-pseudo-random.ts';
-export * from './get-numbers-around.ts';
+export type * from './types.ts';
